fuel_route_api/fuel_route_project/requirements.txt
django>=5.0,<6.1
djangorestframework>=3.15
requests>=2.31
python-dotenv>=1.0
zipcodes>=1.3
